Fix code quality

+++ src/Bridge/FosUser/EventListener.php
@@ -50,14 +50,11 @@ final class EventListener
             return;
         }
 
-        switch ($request->getMethod()) {
-            case 'DELETE':
-                $this->userManager->deleteUser($user);
-                $event->setControllerResult(null);
-                break;
-            default:
-                $this->userManager->updateUser($user);
-                break;
+        if ('DELETE' === $request->getMethod()) {
+            $this->userManager->deleteUser($user);
+            $event->setControllerResult(null);
+        } else {
+            $this->userManager->updateUser($user);
         }
     }
 }
